requirements.txt
alembic==1.13.3
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
flask==3.0.3
Flask-Assets==2.1.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
Flask-Testing==0.8.1
flask-wtf==1.2.1
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing-extensions==4.12.2
urllib3==2.2.3
webassets==2.0
werkzeug==3.0.4
wtforms==3.1.2
zipp==3.20.2

gunicorn==20.1.0

flask-cors==3.0.10
gunicorn